puffy-backend/requirements.txt
pydantic==2.5.3
pydantic-settings==2.1.0
fastapi==0.109.0
sqlalchemy==2.0.30
PyJWT==2.8.0
uvicorn==0.27.0
mysqlclient==2.2.1
python-dotenv==1.0.1
PyNaCl==1.5.0
pytonconnect==0.3.1
redis==5.0.1
sendgrid==6.11.0
pytoniq-core==0.1.39
pycountry==23.12.11
base58==2.1.1
PyYAML==6.0.1
solana==0.34.0
toml==0.10.2
alembic==1.13.1
slowapi==0.1.9

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0
freezegun==1.4.0
